Some code refactoring and minor changes aurusov/raox#295
