+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/dump/ModelStateStorage.java
@@ -1,4 +1,4 @@
-package ru.bmstu.rk9.rao.ui.player;
+package ru.bmstu.rk9.rao.ui.dump;
 
 import java.util.ArrayList;
 import java.util.List;
