+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/dump/Serializer.java
@@ -1,7 +1,5 @@
-package ru.bmstu.rk9.rao.ui.player;
+package ru.bmstu.rk9.rao.ui.dump;
 
-import java.io.BufferedReader;
-import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.Collection;
@@ -48,32 +46,4 @@ public class Serializer {
 
 	}
 
-	public JSONArray retrieveJSONobjectfromJSONfile() {
-		String jsonData = "";
-		BufferedReader br = null;
-		try {
-			String line;
-			br = new BufferedReader(new FileReader("/home/timur/JSON/test.json"));
-			while ((line = br.readLine()) != null) {
-				jsonData += line + "\n";
-			}
-
-		} catch (IOException e) {
-			e.printStackTrace();
-		} finally {
-			try {
-				if (br != null)
-					br.close();
-			} catch (IOException ex) {
-				ex.printStackTrace();
-			}
-
-		}
-		JSONObject obj = new JSONObject(jsonData);
-		JSONArray jsonModelStateObject = new JSONArray();
-		jsonModelStateObject = obj.getJSONArray("Model state array");
-		return jsonModelStateObject;
-		// System.out.println("\n" + jsonModelStateObject);
-
-	}
 }
