+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/dump/Writer.java
@@ -1,10 +1,11 @@
-package ru.bmstu.rk9.rao.ui.player;
+package ru.bmstu.rk9.rao.ui.dump;
 
 import java.util.Arrays;
 import ru.bmstu.rk9.rao.lib.simulator.SimulatorSubscriberManager;
 import ru.bmstu.rk9.rao.lib.simulator.Simulator;
 import ru.bmstu.rk9.rao.lib.simulator.Simulator.ExecutionState;
 import ru.bmstu.rk9.rao.lib.simulator.SimulatorSubscriberManager.SimulatorSubscriberInfo;
+import ru.bmstu.rk9.rao.ui.player.Player;
 import ru.bmstu.rk9.rao.ui.player.Player.PlayingDirection;
 import ru.bmstu.rk9.rao.lib.json.JSONArray;
 import ru.bmstu.rk9.rao.lib.json.JSONObject;
@@ -19,7 +20,7 @@ public class Writer {
 	private final void initializeSubscribers() {
 		simulationSubscriberManager.initialize(
 				Arrays.asList(new SimulatorSubscriberInfo(stateStorageSubscriber, ExecutionState.STATE_CHANGED),
-							  new SimulatorSubscriberInfo(simulationEndSubscriber, ExecutionState.EXECUTION_COMPLETED)));
+						new SimulatorSubscriberInfo(simulationEndSubscriber, ExecutionState.EXECUTION_COMPLETED)));
 
 	}
 
@@ -36,7 +37,7 @@ public class Writer {
 			JSONObject jsonModelStateObject = new JSONObject();
 			jsonModelStateObject.put("Model state array", jsonModelStateArray);
 			serializer.dumpResoursestoJSONfile(jsonModelStateObject);
-			player.PlayerRun(5, 1000, PlayingDirection.BACK);
+			player.Run(5, 1000, PlayingDirection.BACK);
 		}
 
 	}
@@ -44,6 +45,7 @@ public class Writer {
 	public final void deinitializeSubscribers() {
 		simulationSubscriberManager.deinitialize();
 	}
+
 	private final Player player = new Player();
 	private final JSONArray jsonModelStateArray = new JSONArray();
 	public final SimulationEndSubscriber simulationEndSubscriber = new SimulationEndSubscriber();
